invoice item quantity is actually a double

+++ src/main/java/ch/aaap/harvestclient/domain/InvoiceItem.java
@@ -55,7 +55,7 @@ public interface InvoiceItem extends Reference<InvoiceItem> {
     String getDescription();
 
     @Nullable
-    Long getQuantity();
+    Double getQuantity();
 
     @Nullable
     Double getUnitPrice();
