+++ src/test/java/ch/aaap/harvestclient/impl/invoice/InvoicesApiCreateTest.java
@@ -76,13 +76,13 @@ class InvoicesApiCreateTest {
     @Test
     void createWithLineItems() {
 
-        long quantity = 20;
+        double quantity = 20;
         double unitPrice = 30;
         String firstName = "test First";
         Invoice creationInfo = ImmutableInvoice.builder()
                 .client(clientReference)
                 .addInvoiceItem(ImmutableInvoiceItem.builder()
-                        .quantity(10L)
+                        .quantity(10.)
                         .unitPrice(1.)
                         .kind(kind)
                         .build())
